% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbivinvgaus.R
\name{PlotBivInvGaus}
\alias{PlotBivInvGaus}
\title{Density Contour Plot for Bivariate Inverse Gaussian Distribution}
\usage{
PlotBivInvGaus(x, y, u1, u2, l1, l2, r, v)
}
\arguments{
\item{x}{vector defining range of non negative variable x}

\item{y}{vector defining range of non negative variable y}

\item{u1}{mean value of variable x}

\item{u2}{mean value of variable y}

\item{l1}{shape parameter of variable x}

\item{l2}{shape parameter of variable y}

\item{r}{correlation coefficient of variable X and Y}

\item{v}{correlation coefficient of bivariate normal distribution (Z1, Z2)}
}
\value{
Density contour plot for bivariate inverse Gaussian distribution
}
\description{
Density Contour Plot for Bivariate Inverse Gaussian Distribution
}
\examples{
x=seq(1,10,0.2)
y=seq(1,10,0.2)
v=0.3
r=0.5
l1=4
l2=4
u1=3
u2=3
PlotBivInvGaus(x,y,u1,u2,l1,l2,r,v)

}
\references{
Continuous Bivariate Distributions Second Edition by N. Balakrishnan, Chin-Diew Lai
}
